add variable `WG_QUICK_ROUTING_MODE` to wg-quick to force using fwmark for routing
rationale: Using an Endpoint that is part of a network in AllowedIPs
(e.g. Endpoint=162.12.13.1 and AllowedIPs=162.12.13.0/24) will cause a
looping route when enabling a wireguard connection, because of the
routes created by AllowedIPs values. This could be solved using the
more advanced fwmark-based routing (that you describe as "Improved
Rule-based Routing"), but it only takes effect when AllowedIPs contains
0.0.0.0/0. This patch allow users to override that behaviour in a
simply way.

This patch does not fix the root problem, which could be addressed in
one of the following ways:
- Finding Endpoints that are part of networks present in AllowedIPs,
and only enabling fwmark routing for them
- Always use fwmark routing, even if we don't need it
I do not have enough knowledge to choose which option to implement
(although I prefer #2 because of implementation simplicity). I would
like to submit another patch for this, but I will need guidance on what
option to choose.